Refactor columns metadata
