+++ src/components/journeys/JourneyInstancesDataTable/getColumns.tsx
@@ -1,20 +1,22 @@
 import { GridColDef } from '@mui/x-data-grid-pro';
 
 import { getStaticColumns } from './getStaticColumns';
+import getTagColumns from './getTagColumns';
+import { TagMetadata } from 'pages/api/organize/[orgId]/journeys/[journeyId]/getTagMetadata';
 import { ZetkinJourney } from 'types/zetkin';
 
 export type ColumnNames = Record<string, string>;
 
 const getColumns = (
   columnNames: ColumnNames,
-  dynamicColumns: GridColDef[],
+  tagMetadata: TagMetadata,
   journey: ZetkinJourney
 ): GridColDef[] => {
   const staticColumns = getStaticColumns(journey, columnNames);
   return (
     staticColumns
       .splice(0, 2)
-      .concat(dynamicColumns)
+      .concat(getTagColumns(tagMetadata, columnNames))
       // Add/override common props
       .concat(staticColumns)
       .map((col) => ({
