+++ src/pages/organize/[orgId]/journeys/[journeyId]/index.tsx
@@ -1,5 +1,4 @@
 import { GetServerSideProps } from 'next';
-import { GridColDef } from '@mui/x-data-grid-pro';
 import Head from 'next/head';
 
 import AllJourneyInstancesLayout from 'layout/organize/AllJourneyInstancesLayout';
@@ -8,6 +7,7 @@ import getOrg from 'fetching/getOrg';
 import JourneyInstancesDataTable from 'components/journeys/JourneyInstancesDataTable';
 import { PageWithLayout } from 'types';
 import { scaffold } from 'utils/next';
+import { TagMetadata } from 'pages/api/organize/[orgId]/journeys/[journeyId]/getTagMetadata';
 import ZetkinQuery from 'components/ZetkinQuery';
 import { journeyInstancesResource, journeyResource } from 'api/journeys';
 import { ZetkinJourney, ZetkinJourneyInstance } from 'types/zetkin';
@@ -74,8 +74,8 @@ const JourneyInstancesOverviewPage: PageWithLayout<
             journey,
             ...(journeyInstancesQuery.data as {
               columnNames: ColumnNames;
-              dynamicColumns: GridColDef[];
               journeyInstances: ZetkinJourneyInstance[];
+              tagMetadata: TagMetadata;
             }),
           }}
         />
